+++ src/main/java/org/zeith/cfcore4j/schemas/Mod.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public class Mod
 {
-	public final int id;
-	public final int gameId;
+	public final long id;
+	public final long gameId;
 	public final String name;
 	public final String slug;
 	public final ModLinks links;
@@ -17,12 +17,12 @@ public class Mod
 	public final ModStatus status;
 	public final long downloadCount;
 	public final boolean isFeatured;
-	public final int primaryCategoryId;
+	public final long primaryCategoryId;
 	public final List<Category> categories;
 	public final List<ModAuthor> authors;
 	public final Nullable<ModAsset> logo;
 	public final List<ModAsset> screenshots;
-	public final int mainFileId;
+	public final long mainFileId;
 	public final List<File> latestFiles;
 	public final List<FileIndex> latestFilesIndexes;
 	public final Instant dateCreated;
@@ -31,8 +31,8 @@ public class Mod
 
 	public Mod(JSONObject $)
 	{
-		this.id = $.getInt("id");
-		this.gameId = $.getInt("gameId");
+		this.id = $.getLong("id");
+		this.gameId = $.getLong("gameId");
 		this.name = $.getString("name");
 		this.slug = $.getString("slug");
 		this.links = new ModLinks($.getJSONObject("links"));
@@ -40,12 +40,12 @@ public class Mod
 		this.status = ModStatus.fromJson($.getInt("status"));
 		this.downloadCount = $.getLong("downloadCount");
 		this.isFeatured = $.getBoolean("isFeatured");
-		this.primaryCategoryId = $.getInt("primaryCategoryId");
+		this.primaryCategoryId = $.getLong("primaryCategoryId");
 		this.categories = Util.parseList($.getJSONArray("categories"), Category::new);
 		this.authors = Util.parseList($.getJSONArray("authors"), ModAuthor::new);
 		this.logo = new Nullable<>(!$.isNull("logo") ? new ModAsset($.getJSONObject("logo")) : null);
 		this.screenshots = Util.parseList($.getJSONArray("screenshots"), ModAsset::new);
-		this.mainFileId = $.getInt("mainFileId");
+		this.mainFileId = $.getLong("mainFileId");
 		this.latestFiles = Util.parseList($.getJSONArray("latestFiles"), File::new);
 		this.latestFilesIndexes = Util.parseList($.getJSONArray("latestFilesIndexes"), FileIndex::new);
 		this.dateCreated = Util.parseDateTime($.getString("dateCreated"));
